Refactor to remove unnecessary jsx file.
